requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Django==4.0
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
pillow==10.2.0
PyMySQL==1.1.0
scripts==3.0
sqlparse==0.4.4
stua==0.3
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
